Restructure new generic collection types to match repo conventions
Apply the file-layout, naming, and parameter-validation conventions from
CLAUDE.md to the four newly added collection types (IndexedSet, OrderedSet,
RangeDictionary, RangeSet):

- Split RangeDictionary.cs (which previously declared three public types) into
  Range.cs, ValueRange.cs, and RangeDictionary.cs so each file holds exactly
  one public top-level type.
- Rename the single-arg struct RangeDictionary<T> to Range<T>; the struct
  represents a half-open interval, not a dictionary, and reads naturally as
  the element type of RangeSet<T>.
- Promote IndexedSet, RangeSet, and RangeDictionary to partial classes and
  move their nested Enumerator structs into <Base>.Enumerator.cs partial
  files, matching IndexedPriorityQueue.Enumerator.cs and RingBackedCollection.Enumerator.cs.
- Replace inline ArgumentNullException / ArgumentOutOfRangeException throws
  with the ThrowHelper.ThrowIfNull / ThrowIfNegative guards mandated by
  CLAUDE.md, keeping a small number of single-line domain checks inline.
- Switch the enumerator concurrent-mutation messages from hardcoded strings
  to ResourceStrings.InvalidOperation_CollectionModified.
- Flesh out XML documentation to include the missing <exception>, <returns>,
  and parameter-condition descriptions called for by the doc-style guide,
  and add British-English remarks documenting the storage strategy (parallel
  arrays plus open-addressing buckets, no BCL collections used as backing
  storage).
